+++ CHANGELOG.md
@@ -27,6 +27,7 @@ All notable changes to this library are documented in this file.
 ### Bug Fixes
 
 - Correct fontscale calculation.
+- Correct offset calculation in SIGMETs.
 - Correct web links for SPC day 2,3 convective outlook.
 - Correct metadata for a number of NCEI IGRA2 sites.
 - Fix appending `product_id` to IEM Access summary table for DSM parsing.